add fuel  & send to wallet pages

+++ lib/page/home_page/bloc/supernode/gateway/cubit.dart
@@ -50,7 +50,8 @@ class GatewayCubit extends Cubit<GatewayState> {
     //Miner Health
     List<MinerHealthResponse> listMinersHealth = [];
     try {
-      listMinersHealth = await supernodeRepository.gateways.minerHealth({"orgId": orgId});
+      listMinersHealth =
+          await supernodeRepository.gateways.minerHealth({"orgId": orgId});
 
       double avgHealth = 0;
       double avgUptimeHealth = 0;
@@ -64,8 +65,8 @@ class GatewayCubit extends Cubit<GatewayState> {
           avgUptimeHealth = minerHealth.uptimeHealth;
           avgMiningFuelHealth = minerHealth.miningFuelHealth;
         } else {
-          sumMiningFuel += minerHealth.miningFuel;
-          sumMiningFuelMax += minerHealth.miningFuelMax;
+          sumMiningFuel += minerHealth.miningFuel.toDouble();
+          sumMiningFuelMax += minerHealth.miningFuelMax.toDouble();
         }
       }
 
@@ -141,5 +142,4 @@ class GatewayCubit extends Cubit<GatewayState> {
       logger.e('rdelete gateway error', e, s);
     }
   }
-
 }
